// apps/web/src/lib/seo/seoMap.generated.ts
// GENERATED FILE. DO NOT EDIT.
export const SEO_MAP = [
  {
    "jurisdictionKey": "CA",
    "slug": "canada",
    "displayName": "Canada"
  },
  {
    "jurisdictionKey": "FR",
    "slug": "france",
    "displayName": "France"
  },
  {
    "jurisdictionKey": "DE",
    "slug": "germany",
    "displayName": "Germany"
  },
  {
    "jurisdictionKey": "IT",
    "slug": "italy",
    "displayName": "Italy"
  },
  {
    "jurisdictionKey": "JP",
    "slug": "japan",
    "displayName": "Japan"
  },
  {
    "jurisdictionKey": "MX",
    "slug": "mexico",
    "displayName": "Mexico"
  },
  {
    "jurisdictionKey": "NL",
    "slug": "netherlands",
    "displayName": "Netherlands"
  },
  {
    "jurisdictionKey": "PT",
    "slug": "portugal",
    "displayName": "Portugal"
  },
  {
    "jurisdictionKey": "ES",
    "slug": "spain",
    "displayName": "Spain"
  },
  {
    "jurisdictionKey": "CH",
    "slug": "switzerland",
    "displayName": "Switzerland"
  },
  {
    "jurisdictionKey": "TH",
    "slug": "thailand",
    "displayName": "Thailand"
  },
  {
    "jurisdictionKey": "AE",
    "slug": "uae",
    "displayName": "United Arab Emirates"
  },
  {
    "jurisdictionKey": "GB",
    "slug": "united-kingdom",
    "displayName": "United Kingdom"
  },
  {
    "jurisdictionKey": "US-AZ",
    "slug": "us-arizona",
    "displayName": "Arizona, US"
  },
  {
    "jurisdictionKey": "US-CA",
    "slug": "us-california",
    "displayName": "California, US"
  },
  {
    "jurisdictionKey": "US-CO",
    "slug": "us-colorado",
    "displayName": "Colorado, US"
  },
  {
    "jurisdictionKey": "US-FL",
    "slug": "us-florida",
    "displayName": "Florida, US"
  },
  {
    "jurisdictionKey": "US-IL",
    "slug": "us-illinois",
    "displayName": "Illinois, US"
  },
  {
    "jurisdictionKey": "US-MA",
    "slug": "us-massachusetts",
    "displayName": "Massachusetts, US"
  },
  {
    "jurisdictionKey": "US-MI",
    "slug": "us-michigan",
    "displayName": "Michigan, US"
  },
  {
    "jurisdictionKey": "US-NV",
    "slug": "us-nevada",
    "displayName": "Nevada, US"
  },
  {
    "jurisdictionKey": "US-NJ",
    "slug": "us-new-jersey",
    "displayName": "New Jersey, US"
  },
  {
    "jurisdictionKey": "US-NY",
    "slug": "us-new-york",
    "displayName": "New York, US"
  },
  {
    "jurisdictionKey": "US-TX",
    "slug": "us-texas",
    "displayName": "Texas, US"
  },
  {
    "jurisdictionKey": "US-WA",
    "slug": "us-washington",
    "displayName": "Washington, US"
  }
] as const;
export type SeoMapEntry = (typeof SEO_MAP)[number];